# docs/openapi.yaml
openapi: 3.1.0
info:
  title: IMan Investment Platform API
  version: 1.0.0
  summary: Canonical production contract for the public website and administration portal.
  description: |
    JSON field names are camelCase. IDs are UUIDs and timestamps are RFC 3339 UTC instants.
    Public reads expose only published content. `/admin/**` requires a JWT plus the permission
    stated by each operation. A SUPER_ADMIN bypasses individual permission checks.
servers:
  - url: /api
    description: Same-origin deployment
...
security: []
paths:
  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthSession' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /auth/me:
    get:
      tags: [Auth]
      operationId: getCurrentUser
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Current session
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CurrentUser' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /public/portfolio:
    get:
      tags: [Public content]
      operationId: listPublishedPortfolio
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Size' }
        - { $ref: '#/components/parameters/Search' }
        - { name: industry, in: query, schema: { type: string, maxLength: 100 } }
        - { name: featured, in: query, schema: { type: boolean } }
        - { name: sort, in: query, schema: { type: string, default: 'displayOrder,asc', pattern: '^(displayOrder|name|createdAt),(asc|desc)$' } }
      responses:
        '200':
          description: Published companies
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PortfolioCompanyPage' }
  /public/insights:
    get:
      tags: [Public content]
      operationId: listPublishedInsights
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Size' }
        - { $ref: '#/components/parameters/Search' }
        - { name: sort, in: query, schema: { type: string, default: 'publishedAt,desc', pattern: '^(publishedAt|title),(asc|desc)$' } }
      responses:
        '200':
          description: Published insights
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InsightPage' }
  /public/insights/{slug}:
    get:
      tags: [Public content]
      operationId: getPublishedInsight
      parameters: [{ $ref: '#/components/parameters/Slug' }]
      responses:
        '200':
          description: Published insight
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Insight' }
        '404': { $ref: '#/components/responses/NotFound' }
  /public/team:
    get:
      tags: [Public content]
      operationId: listPublishedTeam
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Size' }
      responses:
        '200':
          description: Published team profiles
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TeamMemberPage' }
...
  /public/contact:
    post:
      tags: [Public content]
      operationId: createContactMessage
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ContactMessageCreate' }
      responses:
        '201':
          description: Message accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SubmissionReceipt' }
        '422': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /public/jobs:
    get:
      tags: [Careers]
      operationId: listPublishedJobs
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Size' }
        - { $ref: '#/components/parameters/Search' }
        - { name: department, in: query, schema: { type: string, maxLength: 100 } }
        - { name: location, in: query, schema: { type: string, maxLength: 100 } }
        - { name: type, in: query, schema: { $ref: '#/components/schemas/JobType' } }
        - { name: sort, in: query, schema: { type: string, default: 'publishedAt,desc', pattern: '^(publishedAt|title|department),(asc|desc)$' } }
      responses:
        '200':
          description: Published jobs
          content:
            application/json:
              schema: { $ref: '#/components/schemas/JobPage' }
  /public/jobs/{slug}:
    get:
      tags: [Careers]
      operationId: getPublishedJob
      parameters: [{ $ref: '#/components/parameters/Slug' }]
      responses:
        '200':
          description: Published job
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
        '404': { $ref: '#/components/responses/NotFound' }
  /public/jobs/{jobId}/applications:
    post:
      tags: [Careers]
      operationId: applyForJob
      parameters: [{ $ref: '#/components/parameters/JobId' }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [application, resume]
              properties:
                application: { $ref: '#/components/schemas/JobApplicationCreate' }
                resume: { type: string, format: binary, description: 'PDF, DOC, or DOCX; maximum 10 MiB.' }
                coverLetter: { type: string, format: binary, description: 'Optional PDF, DOC, or DOCX; maximum 10 MiB.' }
            encoding:
              application: { contentType: application/json }
      responses:
        '201':
          description: Application received
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SubmissionReceipt' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /public/business-plans:
    post:
      tags: [Business plan submission]
      operationId: createBusinessPlan
      description: Creates a server-side draft or atomically creates and submits it. The returned draft token is shown only once.
      parameters:
        - { name: submit, in: query, schema: { type: boolean, default: false } }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [plan]
              properties:
                plan: { $ref: '#/components/schemas/BusinessPlanWrite' }
                documents:
                  type: array
                  maxItems: 10
                  items: { type: string, format: binary }
            encoding:
              plan: { contentType: application/json }
      responses:
        '201':
          description: Draft created or submitted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BusinessPlanDraftAccess' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /public/business-plans/{planId}:
    get:
      tags: [Business plan submission]
      operationId: getOwnBusinessPlanDraft
      parameters:
        - { $ref: '#/components/parameters/PlanId' }
        - { $ref: '#/components/parameters/DraftToken' }
      responses:
        '200':
          description: Draft
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BusinessPlan' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Business plan submission]
      operationId: autosaveBusinessPlanDraft
      description: Idempotent autosave of supplied fields. Only DRAFT plans are mutable.
      parameters:
        - { $ref: '#/components/parameters/PlanId' }
        - { $ref: '#/components/parameters/DraftToken' }
        - { $ref: '#/components/parameters/IfMatch' }
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema: { $ref: '#/components/schemas/BusinessPlanPatch' }
      responses:
        '200':
          description: Saved draft
          headers:
            ETag: { schema: { type: string }, description: New entity version. }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BusinessPlan' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived access JWT issued by `/auth/login`.
  parameters:
...
    PlanId: { name: planId, in: path, required: true, schema: { type: string, format: uuid } }
    JobId: { name: jobId, in: path, required: true, schema: { type: string, format: uuid } }
...
    Slug: { name: slug, in: path, required: true, schema: { type: string, minLength: 1, maxLength: 200, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$' } }
    Page: { name: page, in: query, description: Zero-based page index., schema: { type: integer, minimum: 0, default: 0 } }
    Size: { name: size, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
    Search: { name: search, in: query, schema: { type: string, maxLength: 200 } }
...
    DraftToken: { name: X-Draft-Token, in: header, required: true, schema: { type: string, minLength: 32 }, description: Opaque secret returned once when the draft is created; never place it in a URL. }
    IfMatch: { name: If-Match, in: header, required: true, schema: { type: string }, description: ETag from the most recent representation; prevents lost updates. }
...
  responses:
    Unauthorized: { description: 'Missing, invalid, or expired credentials.', content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
    Forbidden: { description: Authenticated but not authorized., content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
    NotFound: { description: Resource not found., content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
    Conflict: { description: State or uniqueness conflict., content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
    PreconditionFailed: { description: ETag is stale., content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
    PayloadTooLarge: { description: Upload exceeds an endpoint or aggregate limit., content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
    UnsupportedMediaType: { description: File type is not allowed., content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } } }
    ValidationError: { description: Request validation failed., content: { application/problem+json: { schema: { $ref: '#/components/schemas/ValidationProblem' } } } }
    RateLimited:
      description: Too many requests.
      headers: { Retry-After: { schema: { type: integer }, description: Seconds until retry. } }
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
  schemas:
    Problem:
      type: object
      required: [type, title, status, detail, instance, requestId]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
        requestId: { type: string }
        code: { type: [string, 'null'], description: Stable machine-readable error code. }
    ValidationProblem:
      allOf:
        - { $ref: '#/components/schemas/Problem' }
        - type: object
          required: [errors]
          properties:
            errors:
              type: array
              items:
                type: object
                required: [field, code, message]
                properties:
                  field: { type: string, description: 'JSON Pointer, e.g. `/founder/email`.' }
                  code: { type: string }
                  message: { type: string }
                  rejectedValue: {}
    PageMeta:
      type: object
      required: [page, size, totalElements, totalPages, first, last, empty, sort]
      properties:
        page: { type: integer, minimum: 0 }
        size: { type: integer, minimum: 1 }
        totalElements: { type: integer, format: int64, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
        first: { type: boolean }
        last: { type: boolean }
        empty: { type: boolean }
        sort: { type: string, example: 'createdAt,desc' }
    AuditFields:
      type: object
      required: [id, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    UserSummary:
      type: object
      required: [id, displayName, email]
      properties:
        id: { type: string, format: uuid }
        displayName: { type: string }
        email: { type: string, format: email }
        avatarUrl: { type: [string, 'null'], format: uri }
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email: { type: string, format: email, maxLength: 200 }
        password: { type: string, format: password, minLength: 12, maxLength: 128 }
    AuthSession:
      type: object
      required: [accessToken, tokenType, expiresAt, user]
      properties:
        accessToken: { type: string, writeOnly: true }
        tokenType: { type: string, const: Bearer }
        expiresAt: { type: string, format: date-time }
        user: { $ref: '#/components/schemas/CurrentUser' }
    CurrentUser:
      allOf:
        - { $ref: '#/components/schemas/UserSummary' }
        - type: object
          required: [firstName, lastName, roles, permissions]
          properties:
            firstName: { type: string }
            lastName: { type: string }
            roles: { type: array, uniqueItems: true, items: { type: string } }
            permissions: { type: array, uniqueItems: true, items: { type: string } }
    ContentStatus: { type: string, enum: [DRAFT, PUBLISHED, ARCHIVED] }
    BusinessPlanStatus: { type: string, enum: [DRAFT, SUBMITTED, REVIEWING, APPROVED, REJECTED, ARCHIVED] }
...
    JobStatus: { type: string, enum: [DRAFT, PUBLISHED, CLOSED, ARCHIVED] }
    JobType: { type: string, enum: [FULL_TIME, PART_TIME, CONTRACT, INTERNSHIP, TEMPORARY] }
    MediaType: { type: string, enum: [IMAGE, DOCUMENT, VIDEO, OTHER] }
...
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount: { type: number, minimum: 0, multipleOf: 0.01 }
        currency: { type: string, pattern: '^[A-Z]{3}$', example: USD }
    FileAsset:
      type: object
      required: [id, originalName, mediaType, mimeType, sizeBytes, downloadUrl, createdAt]
      properties:
        id: { type: string, format: uuid }
        originalName: { type: string, maxLength: 255 }
        mediaType: { $ref: '#/components/schemas/MediaType' }
        mimeType: { type: string }
        sizeBytes: { type: integer, format: int64, minimum: 0 }
        downloadUrl: { type: string, format: uri-reference }
        createdAt: { type: string, format: date-time }
    PortfolioCompanyWrite:
      type: object
      additionalProperties: false
      required: [name, industry, status]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        industry: { type: string, minLength: 1, maxLength: 100 }
        description: { type: [string, 'null'], maxLength: 10000 }
        imageAssetId: { type: [string, 'null'], format: uuid }
        websiteUrl: { type: [string, 'null'], format: uri, maxLength: 500 }
        featured: { type: boolean, default: false }
        displayOrder: { type: integer, minimum: 0, default: 0 }
        status: { $ref: '#/components/schemas/ContentStatus' }
    PortfolioCompany:
      allOf:
        - { $ref: '#/components/schemas/AuditFields' }
        - { $ref: '#/components/schemas/PortfolioCompanyWrite' }
        - type: object
          properties:
            imageUrl: { type: [string, 'null'], format: uri-reference, readOnly: true }
    InsightWrite:
      type: object
      additionalProperties: false
      required: [title, slug, content, status]
      properties:
        title: { type: string, minLength: 1, maxLength: 300 }
        slug: { type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$', maxLength: 200 }
        category: { type: [string, 'null'], maxLength: 100 }
        excerpt: { type: [string, 'null'], maxLength: 1000 }
        content: { type: string, minLength: 1, maxLength: 1000000 }
        imageAssetId: { type: [string, 'null'], format: uuid }
        authorName: { type: [string, 'null'], maxLength: 200 }
        status: { $ref: '#/components/schemas/ContentStatus' }
        publishedAt: { type: [string, 'null'], format: date-time }
        metaTitle: { type: [string, 'null'], maxLength: 200 }
        metaDescription: { type: [string, 'null'], maxLength: 500 }
    Insight:
      allOf:
        - { $ref: '#/components/schemas/AuditFields' }
        - { $ref: '#/components/schemas/InsightWrite' }
        - type: object
          properties:
            imageUrl: { type: [string, 'null'], format: uri-reference, readOnly: true }
    TeamMemberWrite:
      type: object
      additionalProperties: false
      required: [fullName, roleTitle, status]
      properties:
        fullName: { type: string, minLength: 1, maxLength: 200 }
        roleTitle: { type: string, minLength: 1, maxLength: 100 }
        bio: { type: [string, 'null'], maxLength: 10000 }
        imageAssetId: { type: [string, 'null'], format: uuid }
        linkedinUrl: { type: [string, 'null'], format: uri, maxLength: 500 }
        displayOrder: { type: integer, minimum: 0, default: 0 }
        status: { $ref: '#/components/schemas/ContentStatus' }
    TeamMember:
      allOf:
        - { $ref: '#/components/schemas/AuditFields' }
        - { $ref: '#/components/schemas/TeamMemberWrite' }
        - type: object
          properties:
            imageUrl: { type: [string, 'null'], format: uri-reference, readOnly: true }
...
    JobWrite:
      type: object
      additionalProperties: false
      required: [title, slug, type, description, status]
      properties:
        title: { type: string, minLength: 1, maxLength: 200 }
        slug: { type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$', maxLength: 200 }
        department: { type: [string, 'null'], maxLength: 100 }
        location: { type: [string, 'null'], maxLength: 100 }
        type: { $ref: '#/components/schemas/JobType' }
        status: { $ref: '#/components/schemas/JobStatus' }
        summary: { type: [string, 'null'], maxLength: 1000 }
        description: { type: string, minLength: 1, maxLength: 100000 }
        responsibilities: { type: [string, 'null'], maxLength: 100000 }
        requirements: { type: [string, 'null'], maxLength: 100000 }
        benefits: { type: [string, 'null'], maxLength: 100000 }
        experience: { type: [string, 'null'], maxLength: 100 }
        education: { type: [string, 'null'], maxLength: 100 }
        salaryRange:
          oneOf: [{ $ref: '#/components/schemas/SalaryRange' }, { type: 'null' }]
        metaTitle: { type: [string, 'null'], maxLength: 200 }
        metaDescription: { type: [string, 'null'], maxLength: 500 }
        publishedAt: { type: [string, 'null'], format: date-time, readOnly: true }
        closesAt: { type: [string, 'null'], format: date-time }
    SalaryRange:
      type: object
      required: [minimum, maximum, currency]
      properties:
        minimum: { type: number, minimum: 0, multipleOf: 0.01 }
        maximum: { type: number, minimum: 0, multipleOf: 0.01 }
        currency: { type: string, pattern: '^[A-Z]{3}$' }
        period: { type: string, enum: [HOUR, MONTH, YEAR], default: YEAR }
    Job:
      allOf:
        - { $ref: '#/components/schemas/AuditFields' }
        - { $ref: '#/components/schemas/JobWrite' }
    JobApplicationCreate:
      type: object
      additionalProperties: false
      required: [firstName, lastName, email]
      properties:
        firstName: { type: string, minLength: 1, maxLength: 100 }
        lastName: { type: string, minLength: 1, maxLength: 100 }
        email: { type: string, format: email, maxLength: 200 }
        phone: { type: [string, 'null'], maxLength: 50 }
        linkedinUrl: { type: [string, 'null'], format: uri, maxLength: 500 }
        coverLetterText: { type: [string, 'null'], maxLength: 20000 }
    BusinessPlanWrite:
      type: object
      additionalProperties: false
      required: [founder, company, currentStep]
      properties:
        founder: { $ref: '#/components/schemas/Founder' }
        company: { $ref: '#/components/schemas/CompanyProfile' }
        traction: { $ref: '#/components/schemas/Traction' }
        currentStep: { type: integer, minimum: 1, maximum: 10 }
    BusinessPlanPatch:
      description: Same fields as BusinessPlanWrite, all optional; explicit null clears nullable values.
      type: object
      additionalProperties: false
      properties:
        founder: { $ref: '#/components/schemas/FounderPatch' }
        company: { $ref: '#/components/schemas/CompanyProfilePatch' }
        traction: { $ref: '#/components/schemas/TractionPatch' }
        currentStep: { type: integer, minimum: 1, maximum: 10 }
    Founder:
      type: object
      required: [fullName, email]
      properties:
        fullName: { type: string, minLength: 1, maxLength: 200 }
        position: { type: [string, 'null'], maxLength: 100 }
        email: { type: string, format: email, maxLength: 200 }
        phone: { type: [string, 'null'], maxLength: 50 }
        countryCode: { type: [string, 'null'], pattern: '^[A-Z]{2}$' }
        linkedinUrl: { type: [string, 'null'], format: uri, maxLength: 500 }
    FounderPatch:
      type: object
      properties:
        fullName: { type: [string, 'null'], maxLength: 200 }
        position: { type: [string, 'null'], maxLength: 100 }
        email: { type: [string, 'null'], format: email, maxLength: 200 }
        phone: { type: [string, 'null'], maxLength: 50 }
        countryCode: { type: [string, 'null'], pattern: '^[A-Z]{2}$' }
        linkedinUrl: { type: [string, 'null'], format: uri, maxLength: 500 }
    CompanyProfile:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        websiteUrl: { type: [string, 'null'], format: uri, maxLength: 500 }
        industry: { type: [string, 'null'], maxLength: 100 }
        stage: { type: [string, 'null'], maxLength: 50 }
        teamSize: { type: [integer, 'null'], minimum: 0 }
        foundedOn: { type: [string, 'null'], format: date }
        description: { type: [string, 'null'], maxLength: 20000 }
    CompanyProfilePatch:
      type: object
      properties:
        name: { type: [string, 'null'], maxLength: 200 }
        websiteUrl: { type: [string, 'null'], format: uri, maxLength: 500 }
        industry: { type: [string, 'null'], maxLength: 100 }
        stage: { type: [string, 'null'], maxLength: 50 }
        teamSize: { type: [integer, 'null'], minimum: 0 }
        foundedOn: { type: [string, 'null'], format: date }
        description: { type: [string, 'null'], maxLength: 20000 }
    Traction:
      type: object
      properties:
        fundingRequested:
          oneOf: [{ $ref: '#/components/schemas/Money' }, { type: 'null' }]
        annualRevenue:
          oneOf: [{ $ref: '#/components/schemas/Money' }, { type: 'null' }]
        monthlyGrowthPercent: { type: [number, 'null'], minimum: -100, maximum: 100000 }
    TractionPatch: { $ref: '#/components/schemas/Traction' }
    BusinessPlan:
      allOf:
        - { $ref: '#/components/schemas/AuditFields' }
        - { $ref: '#/components/schemas/BusinessPlanWrite' }
        - type: object
          required: [status, version, documents, notes]
          properties:
            status: { $ref: '#/components/schemas/BusinessPlanStatus' }
            version: { type: integer, minimum: 0 }
            submittedAt: { type: [string, 'null'], format: date-time }
            assignedTo:
              oneOf: [{ $ref: '#/components/schemas/UserSummary' }, { type: 'null' }]
            documents: { type: array, items: { $ref: '#/components/schemas/FileAsset' } }
            notes: { type: array, items: { $ref: '#/components/schemas/Note' } }
    BusinessPlanDraftAccess:
      type: object
      required: [plan, draftToken]
      properties:
        plan: { $ref: '#/components/schemas/BusinessPlan' }
        draftToken: { type: string, writeOnly: true, description: Returned only once. Store securely in browser storage. }
    SubmissionReceipt:
      type: object
      required: [id, reference, status, receivedAt]
      properties:
        id: { type: string, format: uuid }
        reference: { type: string }
        status: { type: string }
        receivedAt: { type: string, format: date-time }
...
    Note:
      type: object
      required: [id, author, content, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        author: { $ref: '#/components/schemas/UserSummary' }
        content: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    ContactMessageCreate:
      type: object
      additionalProperties: false
      required: [name, email, message]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        email: { type: string, format: email, maxLength: 200 }
        phone: { type: [string, 'null'], maxLength: 50 }
        subject: { type: [string, 'null'], maxLength: 200 }
        message: { type: string, minLength: 1, maxLength: 10000 }
...
    PortfolioCompanyPage: { allOf: [{ $ref: '#/components/schemas/PageMeta' }, { type: object, required: [content], properties: { content: { type: array, items: { $ref: '#/components/schemas/PortfolioCompany' } } } }] }
    InsightPage: { allOf: [{ $ref: '#/components/schemas/PageMeta' }, { type: object, required: [content], properties: { content: { type: array, items: { $ref: '#/components/schemas/Insight' } } } }] }
    TeamMemberPage: { allOf: [{ $ref: '#/components/schemas/PageMeta' }, { type: object, required: [content], properties: { content: { type: array, items: { $ref: '#/components/schemas/TeamMember' } } } }] }
    JobPage: { allOf: [{ $ref: '#/components/schemas/PageMeta' }, { type: object, required: [content], properties: { content: { type: array, items: { $ref: '#/components/schemas/Job' } } } }] }
...